requirements.txt
openai
python-dotenv
opencv-python